fix: keep city-credit off the database client import path
The feature commit made src/city-credit.ts import src/core.ts (and, through me-around-you.ts and public-changes.ts, again), and core.ts imports the real database client. The world Postgres harness mocks src/db.ts only after its own static imports are linked, so every bearer-authenticated world route reached the real client and answered 500 (14 of 295 integration tests on CI). HANDLE_RE and postgresErrorCode now live in a dependency-free leaf module, re-exported from core.ts so no other importer changes.
